Move App into ui package
